requirements/base.txt
asgiref==3.2.10
certifi==2020.12.5
celery[redis]==5.1.1
chardet==4.0.0
cryptography==3.4.7
factory-boy==3.2.0
Faker==8.8.1
idna==2.10
PyJWT==2.1.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.5
psycopg2-binary==2.8.5
gunicorn==20.0.4
dj-static==0.0.6
sentry-sdk==0.17.4
sendgrid==6.7.0
twilio==6.50.1
Pillow==8.2.0
drf-rw-serializers==1.0.4

# DJANGO------------------------------------------------------------------
Django==3.1.0
django-redis==5.0.0
django-environ==0.4.5
django-material-admin==1.8.5
django-phonenumber-field[phonenumbers]==5.2.0
django-anymail[sendgrid]==8.3
django-extensions==3.1.3
django-json-widget==1.1.1
django-imagekit==4.0.2
django-tinymce==3.3.0
django-money==2.0

# DJANGO REST FRAMEWORK---------------------------------------------------
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djangorestframework-api-key==2.0.0